# internal/apidocs/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway
  description: >
    Multi-tenant API gateway: JWT authentication, RBAC authorization,
    distributed rate limiting, response caching, request validation,
    resilient proxying, and observability in front of downstream services.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local docker-compose environment

tags:
  - name: Health
    description: Liveness/readiness probes. Never authenticated, rate limited, or cached.
  - name: Auth
    description: Login, refresh, logout, and current-user endpoints.
  - name: RBAC
    description: Read-only role and permission catalog.
  - name: Gateway
    description: Proxied routes forwarded to downstream services per config/routes.json.

paths:
  /health:
    get:
      tags: [Health]
      operationId: getHealth
      summary: Liveness probe
      description: Reports the process is running. Does not check dependencies.
      security: []
      responses:
        "200":
          description: Process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

  /ready:
    get:
      tags: [Health]
      operationId: getReady
      summary: Readiness probe
      description: Reports 200 only if Redis and PostgreSQL are both reachable.
      security: []
      responses:
        "200":
          description: Redis and PostgreSQL are both reachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"
        "503":
          description: One or both dependencies are unreachable.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReadyResponse"

  /roles:
    get:
      tags: [RBAC]
      operationId: listRoles
      summary: List roles
      description: Lists every role and its permissions from the role cache.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Roles listed.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoleResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /permissions:
    get:
      tags: [RBAC]
      operationId: listPermissions
      summary: List permissions
      description: Lists every permission from the role cache.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Permissions listed.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PermissionResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/login:
    post:
      tags: [Auth]
      operationId: login
      summary: Log in
      description: Exchanges tenant-scoped email/password credentials for an access/refresh token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid email, password, or tenant slug.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/refresh:
    post:
      tags: [Auth]
      operationId: refreshTokens
      summary: Refresh tokens
      description: Rotates a valid refresh token for a new access/refresh token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
      responses:
        "200":
          description: Refresh succeeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid or expired refresh token.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/logout:
    post:
      tags: [Auth]
      operationId: logout
      summary: Log out
      description: Revokes a refresh token. Idempotent — revoking an unknown or already-revoked token still returns 200.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LogoutRequest"
      responses:
        "200":
          description: Logged out.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: logged out
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Current user
      description: Returns the authenticated caller, refetched for freshness.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/orders/{id}:
    get:
      tags: [Gateway]
      operationId: getOrder
      summary: Get order (proxied)
      description: >
        Proxied to the orders-service upstream (config/routes.json). Subject
        to JWT auth, rate limiting, response caching, and the `orders:read`
        permission.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Order returned by the upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/orders:
    post:
      tags: [Gateway]
      operationId: createOrder
      summary: Create order (proxied)
      description: >
        Proxied to the orders-service upstream (config/routes.json). Subject
        to JWT auth, request validation, and the `orders:create` permission.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created by the upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /api/inventory/{id}:
    get:
      tags: [Gateway]
      operationId: getInventoryItem
      summary: Get inventory item (proxied)
      description: >
        Proxied to the inventory-service upstream (config/routes.json).
        Subject to JWT auth and rate limiting; no specific permission
        required beyond a valid token.
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Inventory item returned by the upstream.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InventoryItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/RateLimited"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    BadRequest:
      description: Malformed request or failed validation.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing, invalid, or expired access token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Authenticated but missing the required permission.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    RateLimited:
      description: Per-tenant rate limit exceeded.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          example: ok

    ReadyResponse:
      type: object
      properties:
        status:
          type: string
          example: ready
        redis:
          type: string
          example: ok
        postgres:
          type: string
          example: ok

    RoleResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        display_name:
          type: string
        description:
          type: string
        permissions:
          type: array
          items:
            type: string
        is_system_role:
          type: boolean

    PermissionResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        resource:
          type: string
        action:
          type: string
        description:
          type: string

    LoginRequest:
      type: object
      required: [email, password, tenant_slug]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
        tenant_slug:
          type: string

    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    LogoutRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        expires_in:
          type: integer
          example: 900
        token_type:
          type: string
          example: Bearer

    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        tenant_id:
          type: string
          format: uuid
        role:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    Order:
      type: object
      properties:
        id:
          type: string
        customer_email:
          type: string
          format: email
        quantity:
          type: integer
        status:
          type: string
        created_at:
          type: string
          format: date-time

    CreateOrderRequest:
      type: object
      required: [customer_email, quantity]
      properties:
        customer_email:
          type: string
          format: email
        quantity:
          type: integer
          minimum: 1

    InventoryItem:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        quantity:
          type: integer